% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPermutation.R
\name{checkPermutation}
\alias{checkPermutation}
\title{Checks if a permutation is a valid one to include in the permutation test.}
\usage{
checkPermutation(experiment, permutation, permutationType = "balanced")
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{permutation}{A vector reflecting a permutation of the samples in the experiment (the original ordering is that which corresponds with the rows of the \code{sampleSheet} data frame in \code{experiment}).}

\item{permutationType}{A string indicating whether to obtain all ("all") permutations or only "balanced" permutations. For instance, in a two treatment vs. two control setup, "balanced" permutations would only include those permutations where exactly one treatment sample is swapped with one control sample. "balanced" is generally recommended.}
}
\description{
Checks if a permutation is a valid one to include in the permutation test.
}
