% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPermPValue.R
\name{findPermPValue}
\alias{findPermPValue}
\title{Helper function for obtaining empirical p-values from the permutation test.}
\usage{
findPermPValue(obsValue, nullValues, effectType = "two-sided", ignoreNAs = F)
}
\arguments{
\item{obsValue}{The observed value (either a beta coefficient or its corresponding t value).}

\item{nullValues}{A vector of null values (either beta coefficients or the corresponding t values) generated by the permutation test.}

\item{effectType}{The type of effect of experimental condition on differential binding to look for, or how the p-value will be computed. Options include \code{"negative"} (p-value is proportion of null effects <= observed effect), \code{"positive"} (p-value is proportion of null effects >= observed effect), or \code{"two-sided"} (p-value is proportion of null effects in absolute value >= observed effect in absolute value). The option chosen here should correspond with how the lower and upper thresholds are chosen in \code{\link{getRegions}}.}

\item{ignoreNAs}{If \code{TRUE}, p-values will be calculated ignoring any NAs and NaNs among the null values. If \code{FALSE} (the default), NAs will be returned for every p-value if NAs or NaNs are present.}
}
\description{
Helper function for obtaining empirical p-values from the permutation test.
}
