% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatExperiment.R
\name{formatExperiment}
\alias{formatExperiment}
\title{Formats the results from getting read counts into an experiment for the permutation test. Also performs error checking.}
\usage{
formatExperiment(
  assay,
  rowRanges,
  sampleSheet,
  blacklist = NULL,
  expCondition,
  block = NULL,
  covariates = NULL
)
}
\arguments{
\item{sampleSheet}{A data frame describing the samples in the experiment. At minimum, one column is needed to identify the experimental condition of each sample. A blocking variable and other covariates may also be included.}

\item{expCondition}{A string indicating the name of the experimental condition variable in \code{sampleSheet}. Differential binding will be assessed across experimental condition. Permutations will be made across experimental condition. Only two groups (e.g. treatment and control) are currently permitted.}

\item{block}{A string indicating the name of the blocking variable in \code{sampleSheet}. A blocking variable can be used to identify matched pairs, if present, or a similar setup. In the permutation test, samples are only permuted within their respective blocks.}

\item{covariates}{A vector of strings indicating the names of any other covariates in \code{sampleSheet} to be controlled for in the GLS regressions.}

\item{experiment}{A RangedSummarizedExperiment object from \code{windowCounts()} in \code{csaw}, or a comparable function. Must have a count assay in \code{assay(experiment)} and the corresponding GRanges in \code{rowRanges(experiment)}.}

\item{blackList}{A GRanges object containing genomic regions to exclude from further analysis.}
}
\description{
Formats the results from getting read counts into an experiment for the permutation test. Also performs error checking.
}
