% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModeratedTStatisticsFromBioVar.R
\name{getModeratedTStatisticsFromBioVar}
\alias{getModeratedTStatisticsFromBioVar}
\title{Obtains the moderated t statistics by shrinking only the biological variance.}
\usage{
getModeratedTStatisticsFromBioVar(experiment, verbose = T)
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{verbose}{Boolean indicating whether or not to print out messages indicating progress.}
}
\description{
Assumes that lm is used rather than gls when calculating standard errors.
}
