% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPermutations.R
\name{getPermutations}
\alias{getPermutations}
\title{Obtains permutations of samples that both cross experimental condition and stay within blocks. This function is not particularly efficient for experiments with many samples. It may stall if the number of permutations requested is too large.}
\usage{
getPermutations(
  experiment,
  permutationType = "balanced",
  numRandomPerm = NA,
  numCores = 1,
  verbose = T
)
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{permutationType}{A string indicating whether to obtain all ("all") permutations, or only "balanced" permutations. For instance, in a two treatment vs. two control setup, "balanced" permutations would only include those permutations where exactly one treatment sample is swapped with one control sample. "balanced" is generally recommended.}

\item{numRandomPerm}{An integer indicating the number of permutations to randomly sample if not "all" permutations or every "balanced" permutation should be returned. If \code{NA} (the default), every requested permutation is returned.}

\item{numCores}{The number of cores to use for computation. If \code{==1}, then computation is done in serial. Otherwise, it is run in parallel. Running in parallel is only useful for experiments with more than a few samples per experimental group.}

\item{verbose}{Boolean indicating whether or not to print out messages indicating progress.}
}
\description{
Obtains permutations of samples that both cross experimental condition and stay within blocks. This function is not particularly efficient for experiments with many samples. It may stall if the number of permutations requested is too large.
}
