% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegions.R
\name{getRegions}
\alias{getRegions}
\title{Finds candidate regions to assess for differential binding.}
\usage{
getRegions(
  experiment,
  lower = 0,
  upper = 1,
  thresholdType = "percentile",
  minWidth = NA,
  maxGap = 1500,
  verbose = T
)
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{lower}{A lower threshold. Regions with treatment effects below this threshold will analyzed for possible differential binding. In general, either \code{lower} and \code{upper} should provide symmetry (equally large effects in either direction are considered) or they should be asymmetrical such that only negative or only positive effects are considered.}

\item{upper}{An upper threshold. Regions with treatment effects above this threshold will analyzed for possible differential binding. In general, either \code{lower} and \code{upper} should provide symmetry (equally large effects in either direction are considered) or they should be asymmetrical such that only negative or only positive effects are considered.}

\item{thresholdType}{Possible values include \code{"percentile"} and \code{"raw"}. If \code{"percentile"}, the \code{lower} and \code{upper} values will be considered to be percentile cutoffs for treatment effect. If \code{"raw"}, they will instead be considered to be raw value cutoffs for treatment effects.}

\item{minWidth}{Candidate regions narrow than this width will be excluded from analysis. It is strongly recommended to make this larger than the bin/width width used for generating read counts. If left \code{NA} (the default), it will be set to to the median width of the \code{ranges} attribute of \code{experiment} plus one base pair.}

\item{maxGap}{Candidate regions for differential binding analysis containing gaps larger than this value (in base pairs) will be split into multiple regions, which will be analyzed separately.}

\item{verbose}{Boolean indicating whether or not to print out messages indicating progress.}
}
\description{
Finds candidate regions to assess for differential binding.
}
