% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleRegionEffect.R
\name{getSingleRegionEffect}
\alias{getSingleRegionEffect}
\title{Gets the treatment effect on read counts for a single region.}
\usage{
getSingleRegionEffect(experiment, region, verbose = T)
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{region}{A single GRanges region.}

\item{verbose}{Boolean indicating whether or not to print out messages indicating progress. If this function is being used in parallel, \code{verbose} is set to \code{FALSE} to avoid issues when using \code{tryCatch()}.}
}
\description{
Gets the treatment effect on read counts for a single region.
}
