% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTest.R
\name{permutationTest}
\alias{permutationTest}
\title{Performs a permutation test to assess for differential binding.}
\usage{
permutationTest(
  experiment,
  permutations = NA,
  effectType,
  includeBetas = F,
  numCores = 1,
  verbose = T
)
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{permutations}{A matrix where each row corresponds to a permutation of the samples (i.e. permutation of the columns in the \code{counts} matrix).}

\item{effectType}{The type of effect of experimental condition on differential binding to look for, or how the p-value will be computed. Options include \code{"negative"} (p-value is proportion of null effects <= observed effect), \code{"positive"} (p-value is proportion of null effects >= observed effect), or \code{"two-sided"} (p-value is proportion of null effects in absolute value >= observed effect in absolute value). The option chosen here should correspond with how the lower and upper thresholds are chosen in \code{\link{getRegions}}.}

\item{includeBetas}{A boolean indicating whether or not to record the raw treatment effect (beta coefficient) and its corresponding p-value and FDR. By default, only considers the t-value of the treatment effect. Will be forced to be between 1 and \code{detectCores()}.}

\item{numCores}{The number of cores to use for computation. If \code{==1}, then computation is done in serial. Otherwise, it is run in parallel.}

\item{verbose}{Boolean indicating whether or not to print out messages indicating progress.}
}
\description{
Performs a permutation test to assess for differential binding.
}
