% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{transformCounts}
\alias{transformCounts}
\title{Transforms the read counts associated with the experiment. Stores the result in a metadata field of the input \code{experiment}.}
\usage{
transformCounts(experiment, FUN = function(x) log(x + 0.5))
}
\arguments{
\item{experiment}{A RangedSummarizedExperiment object describing the experiment.}

\item{FUN}{A function implementing the transformation. Should take in a matrix (i.e. the count assay) and output a matrix of the same dimensions, preserving ordering of rows/columns. Default is to log-transform counts.}
}
\description{
Transforms the read counts associated with the experiment. Stores the result in a metadata field of the input \code{experiment}.
}
